% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_func.R
\name{add_cols_building}
\alias{add_cols_building}
\title{adding col of howold_building and uilding_before_war}
\usage{
add_cols_building(df)
}
\arguments{
\item{df}{A data which has reti data.}
}
\description{
adding col of howold_building and uilding_before_war
}
