% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_func.R
\name{add_cols_landsize}
\alias{add_cols_landsize}
\title{adding col of land_size and huge_land}
\usage{
add_cols_landsize(df, ptn)
}
\arguments{
\item{df}{A data which has reti data.}

\item{ptn}{A pattern string which is considered as huge}
}
\description{
adding col of land_size and huge_land
}
