% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_func.R
\name{add_cols_station}
\alias{add_cols_station}
\title{adding col of howfar_st and howfar_st_category}
\usage{
add_cols_station(df)
}
\arguments{
\item{df}{A data which has reti data.}
}
\description{
adding col of howfar_st and howfar_st_category
}
