% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{conv_jc2ad}
\alias{conv_jc2ad}
\title{Convert Japan Calendar to AD}
\usage{
conv_jc2ad(gengou_year)
}
\arguments{
\item{gengou_year}{gengou is strings in multibite charactor
which represents year name of Japanese Calender}
}
\value{
integer or NA if can't find proper gengou, this function return NA
}
\description{
Convert Japan Calendar to AD
}
