% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{get_FWdata}
\alias{get_FWdata}
\title{Making farm and woods data from real estate transaction-price infomation data}
\usage{
get_FWdata(path, timecount = FALSE)
}
\arguments{
\item{path}{vector of csvfile's path. If you give more than two
paths as vector, each given data are bound. So you will
get just one data.frame type data.}

\item{timecount}{logi type. you can see how much time did this function
take to finish work.}
}
\description{
Makes farm and woods data from csvfiles for real estate
transaction-price data, which are provied by Ministry of Land,
Infrastructure and Transport (MLIT)
}
