% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{get_LBdata}
\alias{get_LBdata}
\title{Making Building and Land data from real estate transaction-price infomation data}
\usage{
get_LBdata(path, timecount = FALSE, kind = "住宅地")
}
\arguments{
\item{path}{vector of csvfile's path. If you give more than two
paths as vector, each given data are bound. So you will
get just one data.frame type data.}

\item{timecount}{logi type. you can see how much time did this function
take to finish work.}

\item{kind}{character which represents kind is kind of land, there 
are four kind of land, residence, urban, factory, unresidence}
}
\description{
Makes Building and Land data from csvfiles for real estate
transaction-price data, which are provied by Ministry of Land,
Infrastructure and Transport (MLIT)
}
