% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{jc2ad_func}
\alias{jc2ad_func}
\title{Convert Japan Calendar to AD sub function}
\usage{
jc2ad_func(gengou, first_year)
}
\arguments{
\item{gengou}{gengou is strings in multibite charactor which represents
year name of Japanese Calender}

\item{first_year}{what AD year a first year of Japanese calendar is}
}
\value{
function jc2ad return convert function which you've set
}
\description{
Convert Japan Calendar to AD sub function
}
