% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_func.R
\name{make_date_col}
\alias{make_date_col}
\title{making date cols from original date col}
\usage{
make_date_col(date_strings)
}
\arguments{
\item{date_strings}{Vector of transaction date string.}
}
\description{
making date cols from original date col
}
