% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_func.R
\name{make_hfs_col}
\alias{make_hfs_col}
\title{making how far station cols from original how far col}
\usage{
make_hfs_col(hfs_strings)
}
\arguments{
\item{hfs_strings}{Vector of how far string.}
}
\description{
making how far station cols from original how far col
}
