% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_csvfile}
\alias{read_csvfile}
\title{Read csvfiles for real estate transaction-price data}
\usage{
read_csvfile(path)
}
\arguments{
\item{path}{vector of csvfile's path. If you give more than two
paths as vector, each given data are bound. So you will
get just one data.frame type data.}
}
\description{
Reads csvfiles for real estate transaction-price data, which are
provied by Ministry of Land, Infrastructure and Transport (MLIT),
and Returns a data.frame type data.
}
