% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_func.R
\name{reti_extract_by_city}
\alias{reti_extract_by_city}
\title{extract data for favorite cities}
\usage{
reti_extract_by_city(df, cities)
}
\arguments{
\item{df}{source data.frame}

\item{cities}{vector of favorite city names}
}
\description{
extract data for favorite cities
}
