% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_func.R
\name{reti_extract_by_prefecture}
\alias{reti_extract_by_prefecture}
\title{extract data for favorite prefectures}
\usage{
reti_extract_by_prefecture(df, prefectures)
}
\arguments{
\item{df}{source data.frame}

\item{prefectures}{vector of favorite prefecture names}
}
\description{
extract data for favorite prefectures
}
