% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_func.R
\name{reti_extract_by_station}
\alias{reti_extract_by_station}
\title{extract data for favorite station}
\usage{
reti_extract_by_station(df, stations)
}
\arguments{
\item{df}{source data.frame}

\item{station}{vector of favorite station names}
}
\description{
extract data for favorite station
}
