% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_func.R
\name{reti_extract_func}
\alias{reti_extract_func}
\title{sub function for extract data}
\usage{
reti_extract_func(df, favorite, index)
}
\arguments{
\item{df}{A reti data}

\item{favorite}{vector of favorite data}

\item{index}{index for df}
}
\description{
sub function for extract data
}
