% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_func.R
\name{reti_filter_by_city}
\alias{reti_filter_by_city}
\title{filter data by favorite cities}
\usage{
reti_filter_by_city(df, cities)
}
\arguments{
\item{df}{A reti data}

\item{cities}{vector of favorite city names}
}
\description{
This function filters data by city names. This matching is prefix search.
And each vector elements is OR condision.
}
