% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_func.R
\name{reti_filter_by_kind}
\alias{reti_filter_by_kind}
\title{filter reti data by kind of land}
\usage{
reti_filter_by_kind(df, kind)
}
\arguments{
\item{df}{A data which was read from reti csvfile.}

\item{kind}{String which contains Four character "R" or "C" or "F" or "U";
each character represents kind of land. There are four kind of land, 
residence, commrercial, factory and unresidence, which represent
"R", "C", "F", "U", respectively.}
}
\description{
This function filters data by kind of land. There are four kind of land; 
residence, commrercial, factory and unresidence.
}
