% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_func.R
\name{reti_filter_by_prefecture}
\alias{reti_filter_by_prefecture}
\title{filter data by favorite prefectures}
\usage{
reti_filter_by_prefecture(df, prefectures)
}
\arguments{
\item{df}{A reti data}

\item{prefectures}{vector of favorite prefecture names}
}
\description{
This function filters data by prefecture names. This matching is
prefix search. And each vector elements is OR condision.
}
