% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{reti_read_base}
\alias{reti_read_base}
\alias{read_csvfile}
\title{Read csvfiles for real estate transaction price data}
\usage{
reti_read_base(path, timecount = TRUE)

read_csvfile(path, timecount = TRUE)
}
\arguments{
\item{path}{Vector of csvfile's path.}

\item{timecount}{A logical scalar. you can see how much time did this
function take to finish the work.}
}
\value{
A tibble data.
}
\description{
Reads csvfiles for real estate transaction price data, which are
provied by Ministry of Land, Infrastructure and Transport (MLIT),
}
