% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jirei.R
\name{reti_read_kjten}
\alias{reti_read_kjten}
\title{Read the jirei data file}
\usage{
reti_read_kjten(file)
}
\arguments{
\item{file}{csv file of jirei data}
}
\value{
data.frame
}
\description{
Read the special jirei data file, which file
someone can allow to use.
}
