% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{reti_read_LO}
\alias{reti_read_LO}
\alias{get_LOdata}
\title{Making Land only data from reti csvfile}
\usage{
reti_read_LO(path, timecount = FALSE, kind = "RCFU")

get_LOdata(path, timecount = FALSE, kind = "RCFU")
}
\arguments{
\item{path}{Vector of csvfile's path.}

\item{timecount}{A logical scalar. you can see how much time did this
function take to finish the work.}

\item{kind}{String which contains Four character "R" or "C" or "F" or "U";
each character represents kind of land. There are four kind of land, 
residence, commrercial, factory and unresidence, which represent
"R", "C", "F", "U", respectively.}
}
\value{
A tibble data.
}
\description{
This function Reads csvfiles for real estate transaction price data,
which are provied by Ministry of Land, Infrastructure and Transport (MLIT),
and  makes land only data.
}
