% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{reti_read_MS}
\alias{reti_read_MS}
\alias{get_Mdata}
\title{Making apartment data from reti csvfile}
\usage{
reti_read_MS(path, timecount = FALSE)

get_Mdata(path, timecount = FALSE)
}
\arguments{
\item{path}{Vector of csvfile's path.}

\item{timecount}{A logical scalar. you can see how much time did this
function take to finish the work.}
}
\value{
A tibble data.
}
\description{
This function Reads csvfiles for real estate transaction price data,
which are provied by Ministry of Land, Infrastructure and Transport (MLIT),
and  makes apartment data.
}
