% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{reti_readBase}
\alias{reti_readBase}
\title{Read csvfiles for real estate transaction price data}
\usage{
reti_readBase(path, timecount = TRUE)
}
\arguments{
\item{path}{Vector of csvfile's path.}

\item{timecount}{A logical scalar. you can see how much time did this 
function take to finish work.}
}
\value{
A tibble data.
}
\description{
Reads csvfiles for real estate transaction price data, which are
provied by Ministry of Land, Infrastructure and Transport (MLIT),
}
\examples{
\dontrun{
osaka <- reti_readBase("csvdata/27_osaka_Prefecture_20132_20181.csv")
}
}
