% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_func.R
\name{subset_with_kind}
\alias{subset_with_kind}
\title{subsetting reti data with kind}
\usage{
subset_with_kind(df, kind)
}
\arguments{
\item{df}{data.frame which is made from csvfiles for real estate
transaction-price data, which are provied by Ministry of Land,
Infrastructure and Transport (MLIT)}

\item{kind}{string which contains Four character "R" or "C" or "F" or "U";
each character represents kind of land. There are four kind of land, 
residence, commrercial, factory and unresidence, which represent
"R", "C", "F", "U", respectively.}
}
\description{
subsetting reti data with kind
}
