% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{get_LBdata}
\alias{get_LBdata}
\title{Making Building and Land data from real estate transaction-price infomation data}
\usage{
get_LBdata(path, timecount = FALSE, kind = "RCFU")
}
\arguments{
\item{path}{vector of csvfile's path. If you give more than two
paths as vector, each given data are bound. So you will
get just one data.frame type data.}

\item{timecount}{logi type. you can see how much time did this function
take to finish work.}

\item{kind}{string which contains Four character "R" or "C" or "F" or "U";
each character represents kind of land. There are four kind of land,
residence, commrercial, factory and unresidence, which represent
"R", "C", "F", "U", respectively.}
}
\description{
Makes Building and Land data from csvfiles for real estate
transaction-price data, which are provied by Ministry of Land,
Infrastructure and Transport (MLIT)
}
